requirements.txt
attrs==15.2.0
BeautifulSoup==3.2.1
blinker==1.3
chardet==2.3.0
click==6.7
configglue==1.1.2
configobj==5.0.6
cryptography==1.2.3
debtags==2.0
defer==1.0.6
dirspec==13.10
Django==1.11.5
django-adminlte2==0.1.6
django-crispy-forms==1.7.0
django-registration-redux==1.8
djangorestframework==3.7.3
duplicity==0.7.6
ecdsa==0.13
enum34==1.1.2
feedparser==5.1.3
Flask==0.12.2
gunicorn==19.7.1
httplib2==0.9.1
idna==2.0
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.9.6
lockfile==0.12.2
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
nemo-emblems==3.4.1
netifaces==0.10.4
oauthlib==1.0.3
oneconf==0.3.9
PAM==0.4.2
paramiko==1.16.0
pbr==3.1.1
pexpect==4.0.1
Pillow==3.1.2
piston-mini-client==0.7.5
protobuf==2.6.1
psycopg2==2.7.3.1
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
pyinotify==0.9.6
PyJWT==1.5.3
pyliblzma==0.5.3
PyOpenGL==3.0.2
pyOpenSSL==0.15.1
pyparted==3.10.7
pyserial==3.0.1
pysmbc==1.0.15.5
pysqlite==1.0.1
python-apt==1.1.0b1
python-debian==0.1.27
python-openid==2.2.5
python-social-auth==0.3.6
python-xlib==0.14
pytz==2017.2
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
requests-oauthlib==0.8.0
rpm-python==4.12.0.1
service-identity==16.0.0
setproctitle==1.1.8
six==1.10.0
social-auth-app-django==2.0.0
social-auth-core==1.5.0
stevedore==1.27.0
Twisted==16.0.0
urlgrabber==3.9.1
urllib3==1.13.1
uTidylib==0.2
vboxapi==1.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
Werkzeug==0.12.2
yum-metadata-parser==1.1.4
zope.interface==4.1.3
